requirements.txt
langchain==0.1.9
panel==1.3.8
pandas==2.2.1
openai==1.12.0
langchain-experimental==0.0.52
plotly==5.19.0
python-dotenv==1.0.1